Renaming existing components to match rule
