+++ app/components/primer/clipboard_copy.ts
@@ -32,7 +32,7 @@ function showCheck(button: HTMLElement) {
 
 const clipboardCopyElementTimers = new WeakMap<HTMLElement, number>()
 
-document.addEventListener('clipboard-copy', function ({target}) {
+document.addEventListener('clipboard-copy', function({target}) {
   if (!(target instanceof HTMLElement)) return
   if (!target.hasAttribute('data-view-component')) return
