+++ app/components/primer/primer.ts
@@ -1,10 +1,10 @@
+import './alpha/modal_dialog'
+import './alpha/toggle_switch'
+import './alpha/tool_tip'
 import './beta/auto_complete/auto_complete'
-import './clipboard_copy_component'
+import './clipboard_copy'
+import './dropdown'
+import './image_crop'
+import './local_time'
 import './tab_container_component'
 import './time_ago_component'
-import './local_time'
-import './image_crop'
-import './dropdown'
-import './alpha/toggle-switch-element'
-import './alpha/tool-tip-element'
-import './alpha/modal-dialog-element'
